Use the label for listings when using SortMethods
The label is only used when no SortMethod was added to the listing,
but as soon as a SortMethod is added the listing would suddenly be
using the title (episode name).

This is problematic because we have listings that show episodes from
different TV shows in the same listing (e.g. Recent items, Continue
watching, Watch later or TV guide). Only showing the title
(episode name) is useless in these listings, we tend to show the
TV show name + the episode title.

It should always be using the label, just like when no SortMethod was
added.

This fixes this issue.

(For reviewers, the only change is replacing "%T" with "%L" so the Label
is used instead of the Title)