Proposal to increase queue slightly
This proposal is based on a change already implemented by Davilla which
also fixes known issues with specific HLS streams.

While these streams have issues, we are in no position to fix those
streams. Other tools behave perfectly fine on these streams as well and
with this small fix the streams behave well in Kodi as well.

For Matrix we are looking into a more extensive fix in the player, but
while we do not expect large changes in the VideoPlayer in Leia to be
accepted and we like to see this fixed in the next release of Leia we
are proposing this here.

This change also includes an improvement for small-memory devices that
was already implemented in OMXplayer for RPi but is useful here as well.